Crate problem119

Problem 119 (Digit power sum)

Solution detail

Given a perfect power, it is easy to check whether it is a power of the sum of its digits - just add up the digits and see!

We can solve this problem, therefore, by iterating over all perfect powers in increasing order. This can be done by using a priority queue. For each power, store its base, its exponent and its value, and pop them off the priority queue in order of value.

Each time a power is popped, store in the queue the next power up with the same base. Additionally, keep track of the first base not yet included in the queue and make sure to add it at an opportune moment - i.e. when we first include a power whose value exceeds the power with the new base.

To avoid storing useless powers that can never be the solution, we can also calculate a rough lower bound on the smallest exponent to consider for each base. For example, we never need to consider if the square of a 3 digit number will be a solution - because the square will have at most 6 digits, the sum of which will be at most 54.

In general, for a base a with d digits, aᵏ will have at most kd digits, meaning that we only need to consider exponents k for which 9kd > 10ᵈ⁻¹, in other words k > 10ᵈ⁻¹ / 9d.

Structs

Power
A structure that represents a perfect power

PowersIterator
An iterator over sequential perfect powers, ignoring those powers which can never be a digit sum perfect power.

Functions

answer
Solve the problem, returning the answer as a String

main

min_viable_exponent
Returns the smallest exponent which, for this base, could form a digit-sum power.

num_digits
Returns the number of digits in the given number

solve
Find the nth number which is the sum of its digits raised to some perfect power.

sum_of_digits
Returns the sum of the digits in the given number